Move shared doc functions out of doc_upsert
