+++ src/cli/doc_get.rs
@@ -3,7 +3,7 @@
 use super::util::convert_json_value_to_nu_value;
 use crate::state::State;
 
-use crate::cli::doc_upsert::{build_batched_kv_items, get_active_cluster_client_cid};
+use crate::cli::doc_common::{build_batched_kv_items, get_active_cluster_client_cid};
 use crate::cli::util::{cluster_identifiers_from, NuValueMap};
 use crate::client::KeyValueRequest;
 use futures::stream::FuturesUnordered;
