+++ src/cli/doc_replace.rs
@@ -1,6 +1,6 @@
 //! The `doc replace` command performs a KV replace operation.
 
-use crate::cli::doc_upsert::run_kv_store_ops;
+use crate::cli::doc_common::run_kv_store_ops;
 use crate::client::KeyValueRequest;
 use crate::state::State;
 use std::sync::{Arc, Mutex};
